% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arrayExpressDownload}
\alias{arrayExpressDownload}
\title{arrayExpressDownload}
\usage{
arrayExpressDownload(datasets, targetDir = getwd(), filePattern = NULL,
  unzip = "internal", overwrite = FALSE)
}
\arguments{
\item{datasets}{the dataset names, for example: c("E-TABM-43", "E-TABM-158")}

\item{targetDir}{the target directory to store the datasets}

\item{filePattern}{the file pattern of the expected data file extracted
from gzipped file}

\item{unzip}{the path to the command to be used in unzip function}

\item{overwrite}{If TRUE, overwrite existing files, otherwise ignore such
files.}
}
\value{
a data frame containing dataset and how many expected data files in
that dataset
}
\description{
The function downloads array express raw data from EBI ftp server based on
datasets user provided. Once the compressed raw data is downloaded,
individual target file will be extracted from compressed raw data. The
dataset/count table will be returned.
}
\examples{
#download three datasets from ArrayExpress website
rootDir<-paste0(dirname(tempdir()), "/DupChecker")
dir.create(rootDir, showWarnings = FALSE)
datatable<-arrayExpressDownload(datasets = c("E-MEXP-3872"), targetDir=rootDir, filePattern="cel$")
}

