% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geoDownload}
\alias{geoDownload}
\title{geoDownload}
\usage{
geoDownload(datasets, targetDir = getwd(), filePattern = NULL,
  tar = "internal", overwrite = FALSE)
}
\arguments{
\item{datasets}{the GEO dataset names, for example: c("GSE14333")}

\item{targetDir}{the target directory to store the datasets}

\item{filePattern}{the file pattern of the expected data file may or may not
extracted from gzipped file, for example: "cel$" for AffyMetrix
CEL files. Default is NULL.}

\item{tar}{the path to the command to be used in untar function}

\item{overwrite}{If TRUE, overwrite existing files, otherwise ignore such
files. The equivalent of unzip -o.}
}
\value{
a data frame containing dataset and how many target files in
        that dataset
}
\description{
The function downloads GEO raw data from ncbi ftp server based on datasets
user provided. Once the compressed raw data is downloaded,
individual gzipped target file will be extracted from compressed raw data,
and individual target file will be extracted from corresponding
gzipped file. The dataset/count table will be returned.
}
\examples{
#download three datasets from GEO website
rootDir<-paste0(dirname(tempdir()), "/DupChecker")
dir.create(rootDir, showWarnings = FALSE)
datatable<-geoDownload(datasets = c("GSE1478"), targetDir=rootDir, filePattern="cel$")
}

