% Generated by roxygen2 (4.0.0): do not edit by hand
\name{geoDownload}
\alias{geoDownload}
\title{geoDownload}
\usage{
geoDownload(datasets, targetDir = getwd(), filePattern = ".CEL$",
  tar = "internal")
}
\arguments{
\item{datasets}{the GEO dataset names, for example: c("GSE14333", "GSE13067", "GSE17538")}

\item{targetDir}{the target directory to store the datasets}

\item{filePattern}{the file pattern of the expected data file extracted from gzipped file}

\item{tar}{the path to the command to be used which is used in untar function}
}
\value{
a data frame containing dataset and how many target files in that dataset
}
\description{
The function downloads GEO raw data from ncbi ftp server based on datasets user provided. Once the compressed raw data is downloaded,
individual gzipped target file will be extracted from compressed raw data, and individual target file will be extracted from corresponding
gzipped file. The dataset/count table will be returned.
}
\examples{
\dontrun{
datatable<-geoDownload(datasets = c("GSE14333", "GSE13067", "GSE17538"), targetDir=getwd())
}
}

